refactor: add types to bridge mock data
